refactor(marketing): modular page-blocks to stop VES merge cascades
Split the mono page-blocks.ts into page-blocks/pages/* modules with
per-page *PageSeed exports and seed auto-discovery. Concurrent VES
page wires no longer share one mega-file or a hand-edited PAGE_SEEDS
array. Gate: validate:page-blocks-modules (phase-1 hard fail).
